+++ code/mobile/components/sport/sets-view/FinishableEditableSetCard.tsx
@@ -1,0 +1,42 @@
+import { Card } from "@/components/ui/Card";
+import { Set } from "@/lib/stores/sport/fe-sets-store";
+import { Check, Pen } from "lucide-react-native";
+import { Text, TouchableOpacity, View } from "react-native";
+
+export function FinishableEditableSetCard(props: {
+  set: Set;
+  numberOfSetsInExercise: number;
+  isMarkedAsFinished: boolean;
+  onPressEdit: () => void;
+  onToggleFinished: () => void;
+  editable: boolean;
+}) {
+  return (
+    <Card className="mb-1">
+      <View
+        className={`flex-1 flex-row justify-evenly items-center py-2 ${
+          props.isMarkedAsFinished ? "bg-green-300" : ""
+        }`}
+      >
+        <Text className="text-md">{props.set.weightsInKg.toFixed(2)}</Text>
+        <Text className="text-md">{props.set.repetitions}</Text>
+        <Text className="text-md">
+          {(
+            props.numberOfSetsInExercise *
+            props.set.repetitions *
+            props.set.weightsInKg
+          ).toFixed(2)}
+        </Text>
+        {props.editable ? (
+          <TouchableOpacity onPress={props.onPressEdit}>
+            <Pen />
+          </TouchableOpacity>
+        ) : (
+          <TouchableOpacity onPress={props.onToggleFinished}>
+            <Check />
+          </TouchableOpacity>
+        )}
+      </View>
+    </Card>
+  );
+}
